Move plugin to daemon/pkg/plugin
