+++ daemon/cluster/controllers/plugin/controller.go
@@ -12,8 +12,8 @@ import (
 	"github.com/docker/docker/api/types/backend"
 	"github.com/docker/docker/api/types/registry"
 	"github.com/docker/docker/api/types/swarm/runtime"
+	"github.com/docker/docker/daemon/pkg/plugin"
 	v2 "github.com/docker/docker/daemon/pkg/plugin/v2"
-	"github.com/docker/docker/plugin"
 	"github.com/gogo/protobuf/proto"
 	"github.com/moby/swarmkit/v2/api"
 	"github.com/pkg/errors"
