+++ daemon/daemon.go
@@ -50,6 +50,7 @@ import (
 	pluginexec "github.com/docker/docker/daemon/internal/plugin/executor/containerd"
 	dlogger "github.com/docker/docker/daemon/logger"
 	"github.com/docker/docker/daemon/network"
+	"github.com/docker/docker/daemon/pkg/plugin"
 	"github.com/docker/docker/daemon/snapshotter"
 	"github.com/docker/docker/daemon/stats"
 	"github.com/docker/docker/distribution"
@@ -68,7 +69,6 @@ import (
 	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/sysinfo"
-	"github.com/docker/docker/plugin"
 	refstore "github.com/docker/docker/reference"
 	"github.com/docker/docker/registry"
 	volumesservice "github.com/docker/docker/volume/service"
