+++ daemon/internal/metrics/plugin_unix.go
@@ -12,9 +12,9 @@ import (
 	"time"
 
 	"github.com/containerd/log"
+	"github.com/docker/docker/daemon/pkg/plugin"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/plugins"
-	"github.com/docker/docker/plugin"
 	gometrics "github.com/docker/go-metrics"
 	"github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/pkg/errors"
