+++ daemon/internal/metrics/plugin_unsupported.go
@@ -3,8 +3,8 @@
 package metrics
 
 import (
+	"github.com/docker/docker/daemon/pkg/plugin"
 	"github.com/docker/docker/pkg/plugingetter"
-	"github.com/docker/docker/plugin"
 )
 
 func RegisterPlugin(*plugin.Store, string) error { return nil }
