+++ daemon/pkg/plugin/defs.go
