+++ daemon/pkg/plugin/errors.go
