+++ daemon/pkg/plugin/events.go
