+++ daemon/pkg/plugin/manager.go
