+++ daemon/pkg/plugin/manager_linux.go
