+++ daemon/pkg/plugin/manager_linux_test.go
