+++ daemon/pkg/plugin/manager_test.go
