+++ daemon/pkg/plugin/manager_windows.go
