+++ daemon/pkg/plugin/progress.go
