+++ daemon/pkg/plugin/registry.go
