+++ daemon/pkg/plugin/store.go
