+++ daemon/pkg/plugin/store_test.go
